reslab/requirements.txt
---
title: requirements.txt
date: 2025-10-16
version: draft
checksum: 95b207a963f3
---

numpy
pandas
matplotlib
pyyaml
tqdm
scipy

## Methods
Briefly state datasets, parameters, seeds, and procedures.

## Limitations
List key caveats (sampling bias, small N, model assumptions).

## Evidence & Links
- [Link 1](#)
- [Link 2](#)

Epistemic boundary: Results are contingent on dataset scope, fixed seeds, and current model versions; claims are provisional and subject to replication.
